the result of the 32bit non-vector sum is returned as u64

+++ include/gather/simd_variants/avx/agg_avx_32BitVariants.h
@@ -12,7 +12,7 @@
  * @return uint64_t
  */
 
-uint32_t aggregate_scalar(const uint32_t* array, uint64_t number, const uint32_t stride=0) {
+uint64_t aggregate_scalar(const uint32_t* array, uint64_t number, const uint32_t stride=0) {
         uint32_t res = 0;
         for (uint64_t i = 0; i < number; i++)
             res += array[i];
